matches list displayed correctly with data

+++ app/src/main/java/com/better/model/dataHolders/Goals.kt
@@ -1,7 +1,7 @@
 package com.better.model.dataHolders
 
 data class Goals(
-    val home: Int?,
-    val away: Int?
+    val home: Long?,
+    val away: Long?
 )
